Moved GlobalData model class out of model.py
