+++ iot_api/user_api/models/__init__.py
@@ -19,3 +19,4 @@ from iot_api.user_api.models.GatewayToTag import GatewayToTag
 from iot_api.user_api.models.NotificationAssetTag import NotificationAssetTag
 from iot_api.user_api.models.AppKey import AppKey
 from iot_api.user_api.models.ChangeEmailRequests import ChangeEmailRequests
+from iot_api.user_api.models.GlobalData import GlobalData
